cetiTagApp: move fpga and iox to device driver
