+++ packages/ceti-tag-data-capture/src/cetiTagApp/device/fpga.h
@@ -13,7 +13,7 @@
 //-----------------------------------------------------------------------------
 // Libraries
 //-----------------------------------------------------------------------------
-#include "utils/error.h"
+#include "../utils/error.h"
 
 #include <stdint.h> //for uint16_t
