+++ packages/ceti-tag-data-capture/src/cetiTagApp/device/i2c.h
@@ -15,9 +15,12 @@
 
 //I2C bus 1
 #define IOX_I2C_BUS          (1)
+#define BMS_I2C_BUS          (1)
 
-
-#define IOX_I2C_DEV_ADDR     (0x21)
+// BMS I2C Device Address
+#define BMS_I2C_DEV_ADDR_UPPER  (0x0b)  // For internal memory range 180h-1FFh
+#define IOX_I2C_DEV_ADDR        (0x21)
+#define BMS_I2C_DEV_ADDR_LOWER  (0x36)  // For internal memory range 000h-0FFh
 
 
 #endif // __CETI_WHALE_TAG_HAL_I2C_H__
