+++ packages/ceti-tag-data-capture/src/cetiTagApp/recovery.c
@@ -8,7 +8,7 @@
 #include "recovery.h"
 
 #include "cetiTag.h"
-#include "iox.h"
+#include "device/iox.h"
 #include "launcher.h" // for g_stopAcquisition, sampling rate, data filepath, and CPU affinity
 #include "systemMonitor.h" // for the global CPU assignment variable to update
 #include "utils/config.h"
